requirements.txt
webdriver_manager==4.0.0
lxml==4.9.3
selenium==4.11.2
SQLAlchemy==2.0.20
PyMySQL==1.1.0
pydantic==2.3.0
pytest==7.4.0
pytz==2023.3.post1
python_dateutil==2.8.2